- Changed logic in increasing par counts of holes (now the changes are only visual until user chooses to save).

+++ app/src/main/java/fi/efelantti/frisbeegolfer/model/Course.kt
@@ -10,7 +10,7 @@ import kotlinx.android.parcel.Parcelize
 // TODO - Figure out how to parcelize
 class Course(
     @PrimaryKey(autoGenerate = true)
-    var courseId: Int = 0,
+    var courseId: Long = 0,
     var name: String? = "",
     var city: String? = ""
 ) : Parcelable {
